Add three_triangles_stack module.
This new test model is a stack of two ThreeTriangles, where the top can be
turned independently, or both layers can be turned simultaneously.  This model
is in preparation for testing a composite move table, where we want two
independent move tables working together to act as one, allowing us to be more
memory efficient then building out a giant move table for the product of both
possible states.