Include self-symmetry table as part of rep table.
This work helps us create composite tables, which need to consider all possible
symmetries that could reduce a position in order to correctly calculate the
global minimum of the two composite positions.  This allows us an optimization
where we can immediately know if a default symmetry is also the only symmetry
that needs to be considered, which will be true for the vast majority of
positions, as they are not usually self-symmetric.