Eliminate Perm from the RepresentativeTable sig.
This change drops the Permutation used from the RepresentativeTable, and only makes it a generic for new, since building the table is the only time we need a Permutation that can actually be acted on, it's all pre-calculation and look-ups after that.  The reality is, that an Index can map to any number of permutations, and it doesn't really matter which one was used to generate the table, the result should always be the same, if the From<Index> is valid and consistent.