Add and use table_traits for Move & PruningTable.
Add a set of table traits that do all of our key MoveTable actions, and implement them for MoveTable.  Then, instead of expecting that PruningTable accept a MoveTable, accept a generic type that implements the needed table traits.  This gets us a step closer to being able to use CompositeMoveTables to construct a pruning table.