Add ThreeTrapezoids test model.
This model is a bit larger than we'd like to work with, as it has a state count in the hundreds, even with symmetry reduciton, which makes it fairly impractical to evaluate correctness lists of permutations by a human eye.  However, this model is useful for working with patterns (partial permutations).  This model can track a) track more that one piece, b) acutally be incomplete, c) support symmetries that aren't their own inverse.  If we tried to test partial permutations on ThreeTriangles, we can't get all three.  Just tracking the center gives us only B & C, tracking the center and right piece give us A & B (but only mirroring is supported, which is its own inverse), tracking any two edge pieces doesn't give us rotational symmetry, and tracking three or more pieces isn't an incomplete permutation as the last piece is always known from the three.  This model seems like the smallest possible that can check all the boxes.