Allow distinct symmetries in CompositePruningTable
Unlike the CompositeMoveTable, the symmetries of CompositePruningTable don't actually interact between the two PruningTables.  It is still important that the two representations don't drift in regards to symmetry (otherwise a turn on one wouldn't match a turn on the other), but we do this by multiplying symmetries as we go, so that the sym-index represents a raw-index that matches exactly what would happen if the turns were applied (something we can otherwise ignore if we just want a lower bound).  The net effect is that neither representation gets translated at all, which means that the translation is totally hidden.  That means they can be distinct for our two different representations.

We expect this will be useful later in phase 1 of our WingEdge solver.  Likely, that will use six (6) pruning tables with four (4) pieces tracked each.  We can align them in a way that they each have four symmetries.  While some of the patterns can share these four symmetries, others can't.  There's no way to arrange six patterns of four in a way that each pattern has the same four symmetries.  Ultimately, without this change, our WingEdge phase 1 tables are 4x times bigger.