Add ThreeTrapezoidsOuter pattern
This is just like ThreeTrapezoidsInner, but it instead tracks the outer ring of pieces.  We add tests for the module itself and also typical tests for basic usage of it in the MoveTable and PruningTables.  What this really enables though is to show how we can decompose permutations that would be too large as move tables into different patterns that still reassemble as as whole, either via CompositeMoveTables or CompositePruningTables.