Use typical permutation representation or add TODO
In all of our modules, we've been using the inverse of the standard representation of permutations.  Both are technically correct, and you can get one by inverting the other.  But the standard one is a bit easier to reason about and it's much better for patterns, as we need to be able to say "piece i moves to position a[i]."  The standard variety can efficiently use direct lookups, but the inversion can't.

We update all of our test models to use the standard approcah, and mark all of our other implementation with TODOs to highlight this otherwise.