Add a TranslatedPruningTable
The translated pruning table allows exploring a pattern as if it were offset by some translation symmetry, and it's most useful when this translation symmetry isn't actually a valid symmetry on the pattern explored, because this lets us create very large composite pruning tables with much less memory.  A permutation cut into two (or more) patterns that are symmetry can be recombined as a composite table to then make a whole.  For example, if we can create a pruning table for all 3x3x3 edges from a pattern of the top edges and the two front edges in the middle slice.  We'd use U2F2 as a translation table to get a pruning table for the untracked pieces, and then use a composite pruning table to join them together.

We also add some tests to this effect using ThreeTrapezoidsEdge, where a single edge pruning table can be made into a full puzzle pruning table, by composing two more translation pruning tables, each offset by a 120 degree rotation.