Add CoordWingEdgesLockedEvensTopSix
This is a pattern (partial permutation) of just the top six even edges, assuming that even pieces are locked into the even orbit.  So only G1WideTurns and U2 symmetry are supported.  Later we'll make BottomSix to pair with it, which can model the full LockedEvens permuation through a CompositeMoveTable.