Add UD & UDSlice patterns on WingEdgesLockedEvens
The UD pattern represents the four U face and four D face even edges, and the UD slice represents the middle 4 even edges, assuming that pieces are locked in an even orbit.  Combined through a CompositeMoveTable or a CompositePruningTable, we can solve the even edges (again, assuming the orbit is locked).

These two patterns are possibly preferred to the original idea of having both TopSix and BottomSix patterns which combined to make the whole.  Both UD and UDSlice have three symmetries instead of just one for the sixes, so a pruning table based on a composite move table will be half the size.

However, the jury is still out on the asymetric split.  A TopSix pruning table can generated in about 6 seconds.  A pruning table built on a UD+UDSlice composite move table takes about 300 seconds.  While UD has much greater pruning power than TopSix individually, a pruning table built only on UD still takes about 190s to build.  So if the use of UD or a full even pruning table is impractical, then two six piece pruning tables make a lot of sense.  However, we may want a variety that distributes pieces more around the cube, so that our composite pruning tables are a bit more entangled, for a better overall result.