Add H1WideTurn
This move set represents the turns that split the wing edges exactly into odds and evens.  So we support this move set all through wing edges and the locked evens permutation and subpatterns.  Notably though, current findings indicate that solving H1 directly on wing edges is infeasible, as it's a massive position.  It is 245,025x larger than a 3x3x3's G1 position.  So we would likely only ever consider using this move set if we were going to do a three+ phase solve.