Add G1aWideTurns
This is an alternate Domino Reduction or G1-like set of generators, where we allow quarter wide turns on the up and down face.  This results in three orbits of 8-pieces, where the middle slice is one large orbit with odd parity.  This is a much harder position to solve than G1WideTurns creates, but we can still do it in under one second on average.  Current testing puts it at about 350ms.